Scrapy-related utilities are moved to the Apify SDK
