+++ templates/python-beautifulsoup/requirements.txt
@@ -1,6 +1,6 @@
 # Add your dependencies here.
 # See https://pip.pypa.io/en/latest/reference/requirements-file-format/
 # for how to format them
-apify ~= 1.2.0
+apify ~= 1.3.0
 beautifulsoup4 ~= 4.12.2
 httpx ~= 0.25.0
