+++ templates/python-empty/requirements.txt
@@ -1,1 +1,1 @@
-apify ~= 1.2.0
+apify ~= 1.3.0
