--- templates/python-scrapy/src/apify/pipelines.py
@@ -1,18 +1,0 @@
-from itemadapter import ItemAdapter
-from scrapy import Item, Spider
-
-from apify import Actor
-
-
-class ActorDatasetPushPipeline:
-    """
-    Used to output the items into the Actor's default dataset.
-
-    Enabled only when the project is run as an Actor.
-    """
-
-    async def process_item(self, item: Item, spider: Spider) -> Item:
-        item_dict = ItemAdapter(item).asdict()
-        Actor.log.debug(f'Pushing item to the dataset: {item_dict}')
-        await Actor.push_data(item_dict)
-        return item
