+++ templates/python-selenium/.gitignore
@@ -20,6 +20,7 @@ __pycache__
 .dmypy.json
 dmypy.json
 .pytest_cache
+.ruff_cache
 
 .scrapy
 *.log
