+++ templates/python-start/.gitignore
@@ -25,6 +25,7 @@ __pycache__
 .dmypy.json
 dmypy.json
 .pytest_cache
+.ruff_cache
 
 .scrapy
 *.log
