+++ wrappers/python-scrapy/{projectFolder}/main.template.py
@@ -13,13 +13,13 @@ def _get_scrapy_settings(max_depth: int) -> Settings:
     """
     settings = get_project_settings()
     # Add our Actor Push Pipeline with the lowest priority
-    settings['ITEM_PIPELINES']['{{apify_module_path}}.pipelines.ActorDatasetPushPipeline'] = 1
+    settings['ITEM_PIPELINES']['apify.scrapy.pipelines.ActorDatasetPushPipeline'] = 1
     # Disable default Retry Middleware
     settings['DOWNLOADER_MIDDLEWARES']['scrapy.downloadermiddlewares.retry.RetryMiddleware'] = None
     # Add our custom Retry Middleware with the top priority
-    settings['DOWNLOADER_MIDDLEWARES']['{{apify_module_path}}.middlewares.ApifyRetryMiddleware'] = 999
+    settings['DOWNLOADER_MIDDLEWARES']['apify.scrapy.middlewares.ApifyRetryMiddleware'] = 999
     # Add our custom Scheduler
-    settings['SCHEDULER'] = '{{apify_module_path}}.scheduler.ApifyScheduler'
+    settings['SCHEDULER'] = 'apify.scrapy.scheduler.ApifyScheduler'
     settings['DEPTH_LIMIT'] = max_depth
     return settings
